train2

+++ trainer.py
@@ -13,7 +13,7 @@ def loaders(dataset_path, val_frac=.2, batch_size=8, image_size=512, doses=dose2
            DataLoader(datasets[1], batch_size=batch_size*2)
 
 def reshape_batch(batch):
-    n = batch.shape[0]/2
+    n = int(batch.shape[0]/2)
     return batch[:n], batch[n:]
 
 def train(loaders, vgg_enc, epochs=1,
